refactor: move public URL/CORS/SSL properties to DomainProperties
Relocate appBaseUrl, apiBaseUrl, resolveApiBaseUrl(), cors, and ssl from
KinoticProperties (kinotic-core) to DomainProperties (kinotic-domain).
These are application-domain concerns — public-facing URLs the SPA renders
into, browser CORS policy, and HTTP TLS — none of which belong in core.

Make KinoticApiGatewayProperties, KinoticPersistenceProperties, and
KinoticGithubProperties extend KinoticDomainProperties so they continue
to pick up the moved settings. PersistenceVerticleFactory and
OpenApiVertxRouterFactory switch their KinoticProperties injection to
KinoticDomainProperties to reach the relocated cors/ssl.

Spring property paths shift to kinotic.domain.*:
  kinotic.appBaseUrl   → kinotic.domain.appBaseUrl
  kinotic.apiBaseUrl   → kinotic.domain.apiBaseUrl
  kinotic.cors.*       → kinotic.domain.cors.*
  kinotic.ssl.*        → kinotic.domain.ssl.*

YAML / Helm values, KinD + Azure terraform, docker-compose env vars, and
the configmap KINOTIC_* env-var names move with them.
