+++ deployment/docker-compose/compose.kinotic-server.yml
@@ -41,7 +41,7 @@ services:
       OTEL_METRICS_EXPORTER: otlp
       OTEL_LOGS_EXPORTER: otlp
       SPRING_PROFILES_ACTIVE: development,compose
-      KINOTIC_APPBASEURL: "http://localhost:9090"
+      KINOTIC_DOMAIN_APPBASEURL: "http://localhost:9090"
       KINOTIC_DOMAIN_EMAIL_ENABLED: "false"
       SPRING_AI_OPENAI_API-KEY: "noop"
       SPRING_AI_OPENAI_BASE_URL: "https://api.x.ai"
