+++ kinotic-persistence/src/main/java/org/kinotic/persistence/api/config/KinoticPersistenceProperties.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.experimental.Accessors;
-import org.kinotic.core.api.config.KinoticProperties;
+import org.kinotic.domain.api.config.KinoticDomainProperties;
 import org.springframework.stereotype.Component;
 
 /**
@@ -16,7 +16,7 @@ import org.springframework.stereotype.Component;
 @Accessors(chain = true)
 @NoArgsConstructor
 @Component
-public class KinoticPersistenceProperties extends KinoticProperties {
+public class KinoticPersistenceProperties extends KinoticDomainProperties {
 
     /**
      * If true, persistence functionality will not be loaded.
