+++ website/content/2.platform/4.organization-management.md
@@ -201,7 +201,7 @@ For now, per-org SSO can be wired manually:
 
 1. Create the `OidcConfiguration` directly in Elasticsearch (POST through the OpenAPI endpoint or via a migration).
 2. Append its id to the org's `oidcConfigurationIds`.
-3. Add the redirect URI `https://<apiBaseUrl>/api/login/callback/<configId>` to the IdP app registration. For same-origin deploys (`kinotic.apiBaseUrl` unset) this falls back to `<appBaseUrl>`; for split-origin deploys (SPA on Static Web Apps, backend on AKS) it must be the backend's hostname so the IdP returns the browser to the kinotic-server pod, not the SPA.
+3. Add the redirect URI `https://<apiBaseUrl>/api/login/callback/<configId>` to the IdP app registration. For same-origin deploys (`kinotic.domain.apiBaseUrl` unset) this falls back to `<appBaseUrl>`; for split-origin deploys (SPA on Static Web Apps, backend on AKS) it must be the backend's hostname so the IdP returns the browser to the kinotic-server pod, not the SPA.
 
 A user who logs in via this path lands at the same `/api/login/callback/:configId` handler — the IdP doesn't care that the configId is org-scoped instead of platform.
